requirements.txt
boto==2.39.0
bz2file==0.98
click==6.2
gensim==0.12.3
httpretty==0.8.10
nltk==3.1
numpy==1.10.4
ordereddict==1.1
protobuf==3.0.0a3
python-dotenv==0.3.0
requests==2.8.1
scikit-learn==0.17
scipy==0.17.0
six==1.10.0
skflow==0.0.1
sklearn==0.0
smart-open==1.3.2
tensorflow==0.6.0
textblob==0.11.0
wheel==0.26.0
